pyproject.toml
[project]
name = "luminari-backend"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Si Young Byun", email = "siyoung.byun@gmail.com"},
]
dependencies = [
    "django>=5.0.6",
    "djangorestframework>=3.15.2",
    "channels[daphne]>=4.1.0",
    "djangorestframework-simplejwt[crypto]>=5.3.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
lint = [
    "black>=24.4.2",
]
hook = [
    "pre-commit>=3.7.1",
]
